+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- CheckInternalLinks - stop various false positive alarms
+
 ## [0.1.0] - 2025-06-07
 
 First major release